fix: correct type of balance

+++ app/models/client.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Float
+from sqlalchemy import Column, Integer, String, ForeignKey, Float, Numeric
 from models.base import Base
 class Client(Base):
     __tablename__ = "clients"
@@ -6,4 +6,4 @@ class Client(Base):
     id = Column(Integer, primary_key=True, index=True)
     first_name = Column(String, nullable=False)
     last_name = Column(String, nullable=False)
-    initial_balance = Column(Float, default=0.0)
+    initial_balance = Column(Numeric(12,2), default=0.0)
